+++ src/release.rs
@@ -1,6 +1,6 @@
 use crate::{arguments::OutputType, helpers, kind, operator, operator::Operator, CliArgs};
 use cached::proc_macro::cached;
-use clap::{ArgGroup, Parser};
+use clap::{ArgGroup, Parser, ValueHint};
 use indexmap::IndexMap;
 use lazy_static::lazy_static;
 use log::{error, info, warn};
@@ -25,7 +25,7 @@ pub enum CliCommandRelease {
     #[clap(alias("desc"))]
     Describe {
         /// Name of the release to describe
-        #[clap(required = true)]
+        #[clap(required = true, value_hint = ValueHint::Other)]
         release: String,
 
         #[clap(short, long, arg_enum, default_value = "text")]
@@ -40,17 +40,17 @@ pub enum CliCommandRelease {
     ))]
     Install {
         /// Name of the release to install
-        #[clap(required = true)]
+        #[clap(required = true, value_hint = ValueHint::Other)]
         release: String,
 
         /// Whitelist of product operators to install.
         /// Mutually exclusive with `--exclude-products`
-        #[clap(short, long)]
+        #[clap(short, long, value_hint = ValueHint::Other)]
         include_products: Vec<String>,
 
         /// Blacklist of product operators to install.
         /// Mutually exclusive with `--include-products`
-        #[clap(short, long)]
+        #[clap(short, long, value_hint = ValueHint::Other)]
         exclude_products: Vec<String>,
 
         /// If specified a local kubernetes cluster consisting of 4 nodes for testing purposes will be created.
@@ -64,7 +64,8 @@ pub enum CliCommandRelease {
         #[clap(
             long,
             default_value = "stackable-data-platform",
-            requires = "kind-cluster"
+            requires = "kind-cluster",
+            value_hint = ValueHint::Other,
         )]
         kind_cluster_name: String,
     },
@@ -72,7 +73,7 @@ pub enum CliCommandRelease {
     #[clap(alias("un"))]
     Uninstall {
         /// Name of the release to uninstall
-        #[clap(required = true)]
+        #[clap(required = true, value_hint = ValueHint::Other)]
         release: String,
     },
 }