+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Support generation of shell completions ([#54](https://github.com/stackabletech/stackablectl/pull/54))
+
 ### Changed
 
 - Update Rust and Go libs ([#46](https://github.com/stackabletech/stackablectl/pull/46))